+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 ### 🐛 Bug fixes
 
+- Fix input queueing when some of the inputs do not produce frames/samples . ([#625](https://github.com/membraneframework/live_compositor/pull/625) by [@wkozyra95](https://github.com/wkozyra95))
+
 ## [v0.2.0](https://github.com/membraneframework/live_compositor/releases/tag/v0.2.0)
 
 Initial release